Revise attributes:
Currently all attributes are 'rw', which - unless you need it - are
undesireable. The reason is so they could be set during 'BUILD', but
using lazy attributes it is unnecessary.

Instead, they are moved so we can make proper usage of lazy defaults
and lazy attributes.

The user-agent is now 'ro', but has a default with the right
configuration. The request object is also 'ro', but has a lazy builder
that uses the 'ua' which should already have a value.

They were also changed to 'ua' and 'req' to allow users to easily
define them in the instance.