Remove the home-made constant implementation for constant.pm:
The constant.pm module creates proper constants (as in, they
constant fold) and we should use that instead.

The problem (which I assume rose the need for the way it was
implemented) is that there was no separation between the constant
definition and the export definition.

I split it up to defining constants and listing them lower down
for export definition.

I could not find anything that uses Druid::Constants so I made no
changes outside the module.