Use JSON::MaybeXS instead of JSON:
While JSON.pm is a core module, it's considerably slow compared to
a C implementation, available in XS. There's JSON::XS, but using it
forces clients to have a C compiler. JSON::XS also has several issues
the author is not inclined to change. Cpanel::JSON::XS is better, but
still requires a C compiler to install.

JSON::MaybeXS makes this optional, supporting each option based on
priority and whether the client has a C compiler or not.